\name{DetMCD-package}
\alias{DetMCD-package}
\docType{package}
\title{ Deterministic and Robust Algorithms for Data Analysis.}
\description{
This packages contains various robust and deterministic algorithms
 for data analysis.
}
\details{
\tabular{ll}{
Package: \tab DetMCD\cr
Type: \tab Package\cr
Version: \tab 0.0.1\cr
Date: \tab 2012-09-19\cr
Depends: \tab matrixStats, pcaPP (>= 1.8-1), robustbase, MASS\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}

Index:
\preformatted{
DetMCD                  Robust and deterministic estimation of location 
                        and scatter.
DetMCD-package           Robust and Deterministic Algothms for Data 
                        Analysis
plot.DetMCD          Diagnostic plots for DetMCD
DetMCD_CS               Internal function for DetMCD
DetMCD_RW               Internal function for DetMCD
DetMCD_SP               Internal function for DetMCD
xtractR_M               Internal function for DetMCD
quanff                  Internal function for DetMCD
}
}
\author{
Kaveh Vakili [aut, cre],

Maintainer: Kaveh Vakili <vakili.kaveh.email@gmail.com>
}
\keyword{package}
